requirements.txt
schedule
google-api-python-client
python-dotenv
